fix: handle rerender multi time
